fix: complete action on item in auto queue before removing it
AutoProcessQueue now reads item first, then executes function and finally  removes the item, instead of popping it out of queue immediately.
This removes a short window where image is being downloaded and is neither in queue. nor in database, and therefore duplicates could be added.
This also ensures that image won't be lost if server shut downs during its download.